+++ crates/libs/core/src/imp/mod.rs
@@ -1,10 +1,8 @@
 mod bindings;
 mod can_into;
 mod com_bindings;
-mod delay_load;
 mod factory_cache;
 mod generic_factory;
-mod heap;
 mod ref_count;
 mod sha1;
 mod waiter;
@@ -13,24 +11,12 @@ mod weak_ref_count;
 pub use bindings::*;
 pub use can_into::*;
 pub use com_bindings::*;
-pub use delay_load::*;
 pub use factory_cache::*;
 pub use generic_factory::*;
-pub use heap::*;
 pub use ref_count::*;
 pub use sha1::*;
 pub use waiter::*;
 pub use weak_ref_count::*;
-
-pub fn wide_trim_end(mut wide: &[u16]) -> &[u16] {
-    while let Some(last) = wide.last() {
-        match last {
-            32 | 9..=13 => wide = &wide[..wide.len() - 1],
-            _ => break,
-        }
-    }
-    wide
-}
 
 #[doc(hidden)]
 #[macro_export]
