+++ crates/tools/bindings/src/core.txt
@@ -28,9 +28,6 @@
     Windows.Win32.System.LibraryLoader.GetProcAddress
     Windows.Win32.System.LibraryLoader.LOAD_LIBRARY_SEARCH_DEFAULT_DIRS
     Windows.Win32.System.LibraryLoader.LoadLibraryExA
-    Windows.Win32.System.Memory.GetProcessHeap
-    Windows.Win32.System.Memory.HeapAlloc
-    Windows.Win32.System.Memory.HeapFree
     Windows.Win32.System.Threading.CreateEventW
     Windows.Win32.System.Threading.SetEvent
     Windows.Win32.System.Threading.WaitForSingleObject
